refactor: перенести проверки и логирование на уровень сервиса

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -3,9 +3,11 @@ package ru.yandex.practicum.filmorate.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.dao.FriendsDao;
+import ru.yandex.practicum.filmorate.dao.storageimpl.UserDbStorage;
+import ru.yandex.practicum.filmorate.exception.EntityNotFoundException;
+import ru.yandex.practicum.filmorate.exception.IncorrectArgumentsException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.UserStorage;
 
@@ -21,35 +23,77 @@ public class UserService {
     private final FriendsDao friendsDao;
 
     public List<User> findAll() {
+        log.info("Список пользователей получен");
         return userStorage.getAll();
     }
 
-    public User findById(Integer id) {
-
-        return userStorage.getById(id);
+    public User findById(int id) {
+        if (userStorage.contains(id)) {
+            log.info("Пользователь с id={} получен", id);
+            return userStorage.getById(id);
+        } else {
+            throw new EntityNotFoundException(UserDbStorage.class);
+        }
     }
 
-    public List<User> findAllFriends(Integer id) {
-        return userStorage.getAllFriends(id);
+    public List<User> findAllFriends(int id) {
+        if (userStorage.contains(id)) {
+            log.info("Список друзей пользователя c id={} получен", id);
+            return userStorage.getAllFriends(id);
+        } else {
+            throw new EntityNotFoundException(UserDbStorage.class);
+        }
     }
 
-    public List<User> findMutualFriends(Integer idUser, Integer otherIdUser) {
+    public List<User> findMutualFriends(int idUser, int otherIdUser) {
+        if (idUser == otherIdUser) {
+            throw new IncorrectArgumentsException(UserService.class);
+        }
+        if (!userStorage.contains(idUser)) {
+            throw new EntityNotFoundException(UserService.class);
+        }
+        if (!userStorage.contains(otherIdUser)) {
+            throw new EntityNotFoundException(UserService.class);
+        }
+        log.info("Список общих друзей пользователей c id={} и c id={} получен", idUser, otherIdUser);
         return userStorage.getMutualFriends(idUser, otherIdUser);
     }
 
-    public ResponseEntity<?> addToFriends(Integer idUser, Integer idFriend) {
+    public void addToFriends(int idUser, int idFriend) {
+        if (!userStorage.contains(idUser) || !userStorage.contains(idFriend)) {
+            throw new EntityNotFoundException(UserService.class);
+        }
 
-        return friendsDao.addToFriends(idUser, idFriend);
+        int result = friendsDao.addToFriends(idUser, idFriend);
+
+        if (idUser == idFriend || result == 0) {
+            throw new IncorrectArgumentsException(UserService.class);
+        }
+
+        log.info("Пользователь c id={} успешно добавлен в друзья пользователю c id={}", idFriend, idUser);
     }
 
-    public ResponseEntity<?> removeToFriends(Integer idUser, Integer idFriend) {
-        return friendsDao.removeToFriends(idUser, idFriend);
+    public void removeToFriends(int idUser, int idFriend) {
+
+        if (!userStorage.contains(idUser) || !userStorage.contains(idFriend)) {
+            throw new EntityNotFoundException(UserService.class);
+        }
+
+        int result = friendsDao.removeToFriends(idUser, idFriend);
+
+        if (idUser == idFriend || result == 0) {
+            throw new IncorrectArgumentsException(UserService.class);
+        }
+
+        log.info("Пользователь c id={} успешно удален из друзей пользователя c id={}", idFriend, idUser);
+
     }
 
     public User create(User user) {
         if (StringUtils.isEmpty(user.getName())) {
             user.setName(user.getLogin());
         }
+        log.info("Пользователь с id={} добавлен", user.getId());
         return userStorage.create(user);
     }
 
@@ -57,6 +101,13 @@ public class UserService {
         if (StringUtils.isEmpty(user.getName())) {
             user.setName(user.getLogin());
         }
-        return userStorage.upDate(user);
+
+        int result = userStorage.upDate(user);
+
+        if (result == 0) {
+            throw new EntityNotFoundException(UserService.class);
+        }
+        log.info("Пользователь c id={} успешно изменен", user.getId());
+        return user;
     }
 }
